# contracts/openapi/inspector-api.v1.yaml
openapi: 3.0.3
info:
  title: Инспектор ИИ — REST API
  version: 0.24.0
  description: |
    Публичный REST API сервиса «Инспектор ИИ» (веб-клиент и внешние ИС, целевая — ИАИС «РиН»).
    Асинхронная pull-модель: загрузка → process_id → опрос статуса → получение протокола.

    Требования и их ID — docs/TZ.md.

    Соглашения:
    - все поля snake_case; идентификаторы — UUID (кроме params.id — integer, как в ТЗ);
    - время — RFC 3339 (UTC);
    - bbox — [x0, y0, x1, y1] в долях [0;1] видимой страницы (после CropBox/Rotate),
      начало координат — левый верхний угол (как viewport pdf.js);
    - страницы нумеруются с 1;
    - ошибки — объект Error.

    Изменения 0.2.0: пары страниц для экрана сравнения (PagePair, GET /protocols/{id}/page-pairs),
    compliance_percent в сводке, auto_start при загрузке, метод гипотез VISUAL_DIFF, фильтры findings по file_id/evidence_page,
    Finding.page_pair_id и rationale_source.

    Изменения 0.3.0: манифест комплекта при загрузке
    (UploadManifest), external_id объекта, серверный импорт папки (POST /documents/import),
    правка доказательств новой версией (POST /findings/{id}/evidence, evidence_history),
    массовое решение CONFIRM/CLARIFY по одному параметру (POST /findings/bulk-decision),
    PATCH /files — только выбор авторитетной редакции (метаданные не редактируются),
    сводка комплектности (ProcessInfo.completeness).

    Изменения 0.4.0 (Приложение 1 и «Перечень ИД» ред. 1.1): коды параметров M-001…M-132;
...

...
  license:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/processes/{process_id}/finalize:
    parameters:
      - $ref: '#/components/parameters/ProcessId'
    post:
      tags: [processes]
      operationId: finalizeProcess
      summary: Финализировать протокол — кнопка «Завершить» (REQ-VER-07, REQ-VER-08)
      description: |
        Разрешено только когда нет findings в статусе CANDIDATE с inspector_status=PENDING.
        После финализации создаётся запись в integration_outbox для передачи в РиН.
      responses:
        '200':
          description: Финализирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInfo'
        '409':
          description: Есть необработанные кандидаты или неверный статус
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/processes/{process_id}/unfinalize:
    parameters:
      - $ref: '#/components/parameters/ProcessId'
    post:
      tags: [processes]
      operationId: unfinalizeProcess
      summary: Отмена финализации — только ADMIN/SUPERVISOR (REQ-VER-09)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Возвращён в COMPLETED (VERIFICATION_COMPLETED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInfo'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/processes/{process_id}/unfinalize-request:
    parameters:
      - $ref: '#/components/parameters/ProcessId'
    post:
      tags: [processes]
      operationId: requestUnfinalize
      summary: Попросить откат финализации (INSPECTOR, SUPERVISOR) — с причиной, администратору уходит уведомление
      description: |
        Отменить финализацию может только ADMIN или SUPERVISOR (`/unfinalize`). Инспектор, которому нужно
        вернуть проверку в работу, оставляет запрос с причиной; ADMIN и SUPERVISOR получают уведомление
        `UNFINALIZE_REQUESTED`. Открытый запрос на процесс — один (повтор → 409). Успешный `/unfinalize`
        закрывает открытые запросы процесса как `DONE`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '201':
          description: Запрос создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnfinalizeRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/unfinalize-requests:
    get:
      tags: [processes]
      operationId: listUnfinalizeRequests
      summary: Запросы на откат финализации (ADMIN, SUPERVISOR)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UnfinalizeRequestStatus'
      responses:
        '200':
          description: Новые сверху
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnfinalizeRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/unfinalize-requests/{request_id}/reject:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [processes]
      operationId: rejectUnfinalizeRequest
      summary: Отклонить запрос на откат с причиной (ADMIN, SUPERVISOR)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Запрос отклонён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnfinalizeRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/processes/{process_id}/protocols:
    parameters:
      - $ref: '#/components/parameters/ProcessId'
    get:
      tags: [protocols]
      operationId: listProtocolVersions
      summary: История версий протокола (REQ-CMP-11)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProtocolVersionInfo'
  /api/v1/protocols/{protocol_id}:
    parameters:
      - $ref: '#/components/parameters/ProtocolId'
    get:
      tags: [protocols]
      operationId: getProtocol
      summary: Протокол целиком (REQ-CMP-08)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Protocol'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/protocols/{protocol_id}/findings:
    parameters:
      - $ref: '#/components/parameters/ProtocolId'
    get:
      tags: [findings]
      operationId: listFindings
      summary: Findings протокола с фильтрами (очередь верификации)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: finding_status
          in: query
          schema:
            $ref: '#/components/schemas/FindingStatus'
        - name: inspector_status
          in: query
          schema:
            $ref: '#/components/schemas/InspectorStatus'
        - name: section
          in: query
          schema:
            type: string
        - name: review_priority
          in: query
          schema:
            $ref: '#/components/schemas/ReviewPriority'
        - name: param_code
          in: query
          schema:
            type: string
        - name: file_id
          in: query
          description: Только findings с доказательством в этом файле
          schema:
            type: string
            format: uuid
        - name: evidence_page
          in: query
          description: Вместе с file_id — только findings с доказательством на этой странице
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingPage'
  /api/v1/protocols/{protocol_id}/page-pairs:
    parameters:
      - $ref: '#/components/parameters/ProtocolId'
    get:
      tags: [protocols]
      operationId: listPagePairs
      summary: Пары сопоставленных страниц (ПД ↔ РД ↔ ИД) для экрана «Сравнение PDF-схем»
      description: |
        Каждая пара — две страницы разных стадий (или редакций), которые ML счёл одним и тем же листом,
        с матрицей совмещения и областями визуальных различий. Сортировка — по match_score убыв.
      parameters:
        - name: file_id
          in: query
          description: Только пары, где участвует этот файл
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PagePair'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/protocols/{protocol_id}/export:
    parameters:
      - $ref: '#/components/parameters/ProtocolId'
    get:
      tags: [protocols]
      operationId: exportProtocol
      summary: Экспорт протокола в PDF / DOCX / XML / JSON / GOLD / формат ответа организаторов (модуль 7)
      description: |
        Отражает текущее состояние решений инспектора. Для финализированного протокола файл сохраняется
        в хранилище (protocols/{process_id}/v{n}-{время финализации}.{ext}) и повторно не формируется.
      parameters:
        - name: format
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ExportFormat'
      responses:
        '200':
          description: Файл протокола
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Protocol'
                  - $ref: '#/components/schemas/GoldExport'
                  - $ref: '#/components/schemas/SubmissionExport'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/findings/{finding_id}:
    parameters:
      - $ref: '#/components/parameters/FindingId'
    get:
      tags: [findings]
      operationId: getFinding
      summary: Карточка доказательства (REQ-CMP-09, REQ-VER-01)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/findings/{finding_id}/decision:
    parameters:
      - $ref: '#/components/parameters/FindingId'
    post:
      tags: [findings]
      operationId: decideFinding
      summary: Решение инспектора по finding (REQ-VER-02..04)
      description: |
        CONFIRM → CONFIRMED_VIOLATION; REJECT → NEGATIVE_VERIFIED (reason_code и comment обязательны);
        CLARIFY → CLARIFICATION_REQUIRED. Запрещено после финализации (409).
        Пишет audit_log и черновик dataset_items (REQ-ML-02).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: Решение принято
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/findings/{finding_id}/split:
    parameters:
      - $ref: '#/components/parameters/FindingId'
    post:
      tags: [findings]
      operationId: splitFinding
      summary: Разделить составной кандидат на атомарные findings (REQ-VER-05)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitRequest'
      responses:
        '200':
          description: Созданные атомарные findings (исходный помечается как разделённый)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Finding'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/findings/{finding_id}/evidence:
    parameters:
      - $ref: '#/components/parameters/FindingId'
    post:
      tags: [findings]
      operationId: editFindingEvidence
      summary: Правка доказательств инспектором — новая версия
      description: |
        Добавить/удалить доказательные фрагменты. Машинный результат не перезаписывается:
        создаётся новая версия доказательной группы (кто, когда, причина, ссылка),
        предыдущие версии доступны в evidence_history.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceEditRequest'
      responses:
        '200':
          description: Finding с новой текущей версией доказательств
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/findings/bulk-decision:
    post:
      tags: [findings]
      operationId: bulkDecideFindings
      summary: Массовое решение по кандидатам одного параметра (только CONFIRM и CLARIFY)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDecisionRequest'
      responses:
        '200':
          description: Обновлённые findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Finding'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/processes/{process_id}/suspicions:
    parameters:
      - $ref: '#/components/parameters/ProcessId'
    get:
      tags: [suspicions]
      operationId: listSuspicions
      summary: Гипотезы свободного поиска (REQ-SUS)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Suspicion'
  /api/v1/suspicions/{suspicion_id}/decision:
    parameters:
      - $ref: '#/components/parameters/SuspicionId'
    post:
      tags: [suspicions]
      operationId: decideSuspicion
      summary: Отклонить гипотезу или пометить на уточнение
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuspicionDecisionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Suspicion'
  /api/v1/suspicions/{suspicion_id}/promote:
    parameters:
      - $ref: '#/components/parameters/SuspicionId'
    post:
      tags: [suspicions]
      operationId: promoteSuspicion
      summary: Перевести гипотезу в CANDIDATE после привязки доказательств
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoteSuspicionRequest'
      responses:
        '200':
          description: Созданный кандидат
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mtls:
      type: http
      scheme: bearer
      description: >-
        Для внешних ИС — взаимная TLS-аутентификация клиентским сертификатом (УКЭП, REQ-INT-02).
        OpenAPI 3.0 не имеет типа mutualTLS, поэтому схема описана текстом.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    ProcessId:
      name: process_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ProtocolId:
      name: protocol_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Некорректный запрос (валидация OpenAPI)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Операция недопустима в текущем статусе
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    InspectorStatus:
      type: string
      enum: [PENDING, CONFIRMED_VIOLATION, NEGATIVE_VERIFIED, CLARIFICATION_REQUIRED]
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Машиночитаемый код (например UNSUPPORTED_FORMAT, INVALID_STATUS)
        message:
          type: string
          description: Сообщение для пользователя (рус.)
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
...
    UnfinalizeRequestStatus:
      type: string
      enum: [OPEN, DONE, REJECTED]
...
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 3
...
    ProcessInfo:
      allOf:
        - $ref: '#/components/schemas/ProcessStatusInfo'
        - type: object
          required: [object_id, upload_status, created_at]
          properties:
            object_id:
              type: string
              format: uuid
            scenario:
              type: string
              allOf:
                - $ref: '#/components/schemas/CheckScenario'
              nullable: true
            upload_status:
              type: array
              items:
                $ref: '#/components/schemas/StageUploadStatus'
            can_upload:
              type: boolean
            can_verify:
              type: boolean
            can_finalize:
              type: boolean
            counts:
              $ref: '#/components/schemas/FindingCounts'
            completeness:
              $ref: '#/components/schemas/CompletenessSummary'
            created_by:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            finalized_at:
              type: string
              format: date-time
              nullable: true
            finalized_by:
              type: string
              format: uuid
              nullable: true
...